+++ examples/mobc/src/src_user/tlm_cmd/telemetry_definitions.h
@@ -4,7 +4,7 @@
  * @note  このコードは自動生成されています！
  * @note  コード生成元 tlm-cmd-db:
  *          repository:    github.com/arkedge/c2a-core.git
- *          CSV files MD5: 9ca043940fe69f2ae005374c1d34781b
+ *          CSV files MD5: d6e527942b35515e9f81fb3df4eb7f6e
  * @note  コード生成パラメータ:
  *          db_prefix:             SAMPLE_MOBC
  *          tlm_id_range:          [0x00, 0x100]
@@ -30,6 +30,7 @@ typedef enum
   Tlm_CODE_TF = 0x26,
   Tlm_CODE_CA = 0x27,
   Tlm_CODE_DCU = 0x2a,
+  Tlm_CODE_CDRV_UTIL = 0x2d,
   Tlm_CODE_MM = 0x40,
   Tlm_CODE_AM = 0x45,
   Tlm_CODE_APP_TIME = 0x46,